fix(collector): GDN collectors measure serving-parity state, cases cover TP-local geometries
Serving parity (vllm/sglang collect_gdn):
- vLLM context conv precomputes GDNAttentionMetadata outside the timed
  span and passes it to causal_conv1d_fn, matching serving
  (vllm/v1/attention/backends/gdn_attn.py:394-399, utils.py:810-856
  @0.24.0). The metadata=None fallback ran a D2H sync + numpy repacking
  inside every timed call, a constant ~0.1-0.3 ms/call host floor
  (~9 ms/step at 30 GDN layers) that poisoned prior conv rows.
- vLLM recurrent state bf16 -> fp32: Qwen3.5 checkpoints pin
  mamba_ssm_dtype=float32, adopted by verify_and_update_config
  (vllm/model_executor/models/config.py:603-628 @0.24.0); state IO
  dominates decode, so bf16 rows under-measured ~2x at large batch.
  Qwen3.5-only contract, noted at both sites.
- sglang dt_bias fp32 -> model dtype (python/sglang/srt/models/
  qwen3_5.py:237-239 @0.5.14); negligible latency impact, input parity.
- Existing vllm/0.24.0 gdn_perf tables were collected with the old
  harness and must be re-collected before mixing with new data.

TP-local cases (case_generator + model yamls):
- GDN model rows declare tensor_parallel_sizes [1,2,4,8]; the generator
  expands TP-local head counts with a divisibility guard that raises on
  invalid shards, deduplicating on the loader key (16 -> 64 cases per
  backend). Supplies exact-hit grid points for the SDK TP-local GDN
  lookup (0f953cd1); previous global geometries remain the tp=1 subset,
  so checkpoints and existing rows stay valid.
- case_generator.py is in the provenance SHARED_CORE closure, so this
  commit flags every (framework, family) as changed for the evidence
  gate — expected over-inclusive behavior.

+++ collector/sglang/collect_gdn.py
@@ -396,7 +396,7 @@ def run_gdn_generation_benchmark(
             a = torch.randn(batch_size, num_v_heads, dtype=dtype, device=device)
             b = torch.randn(batch_size, num_v_heads, dtype=dtype, device=device)
             a_log = torch.zeros(num_v_heads, dtype=torch.float32, device=device)
-            dt_bias = torch.ones(num_v_heads, dtype=torch.float32, device=device)
+            dt_bias = torch.ones(num_v_heads, dtype=dtype, device=device)
             output = torch.empty(
                 batch_size,
                 1,
